+++ ChangeLog.md
@@ -1,5 +1,8 @@
 ## change log ##
 
+* 2016-04-14: 0.3.1:
+  - start using ui-select
+
 * 2016-04-12: 0.3.1:
   - complete basic sequence to register new ontology in re-hosted mode,
     starting with initial upload of the file then continuing with some